+++ src/main/java/com/example/iesiback/entities/Aporte.java
@@ -19,7 +19,7 @@ public class Aporte {
     private Legajo aporteLegajo;
 
     @Column(name = "aporte_nro_recibo")
-    private Integer aporteNroRecibo;
+    private Long aporteNroRecibo;
 
     @Column(name = "aporte_talonario_recibo")
     private Integer aporteTalonarioRecibo;
@@ -65,11 +65,11 @@ public class Aporte {
         this.aporteMonto = aporteMonto;
     }
 
-    public Integer getAporteNroRecibo() {
+    public Long  getAporteNroRecibo() {
         return aporteNroRecibo;
     }
 
-    public void setAporteNroRecibo(Integer aporteNroRecibo) {
+    public void setAporteNroRecibo(Long  aporteNroRecibo) {
         this.aporteNroRecibo = aporteNroRecibo;
     }
